scheduler/requirements.txt
requests==2.24.0
beautifulsoup4==4.9.2
attrs==20.2.0
pymongo==3.11.0